+++ demos/src/bin/platformer_3d.rs
@@ -123,6 +123,10 @@ fn main() {
                 "CompoundColliders",
                 tnua_demos_crate::levels_setup::compound_colliders_3d::setup_level,
             )
+            .with(
+                "DynamicBodies",
+                tnua_demos_crate::levels_setup::dynamic_bodies_3d::setup_level,
+            )
     });
     app.add_systems(Startup, setup_player);
     app.add_systems(